Fix #994, #996: the census reads the second spelling, seven sites derive their type, and the gates test the Julia the package declares
`Test` and `Docs` were red at e3b2b60ed2 for two causes with nothing in common, and the
separation is what this commit is. Four failures are defects the library owns. Four more
are one toolchain move, and #998 holds the decision they ask for.

**#996, the seven `float(eltype(...))` sites #977 landed.** Each reads its type off the
operation that widens it now. Five seed a Welford accumulator whose division is by an
`Int` count, so the type is `typeof(zero(eltype(x)) / one(Int))`. `coverage_divide` reads
the division it performs, `typeof(zero(eltype(M)) / one(eltype(nu)))`, and
`coverage_frame` reads `eltype(A)`, which the fold that produced `A` already derived.

`NaN` needs a float type and that re-spelling it would widen a `Float32` to `Float64`.
Neither spelling does: `zero(Float32) / one(Int)` is `Float32`. What they do change is an
**exact** element type. A `Rational` accumulator cannot hold the `NaN` sentinel, so it now
raises `InexactError` at the first cell the policy refuses instead of silently reading out
as `Float64`, which is the harm the census exists to stop. Both docstrings say so rather
than guarding it.

**#994, the spelling the census could not see.** `test_55` gains `NAMED_FLOAT`, which
reads a concrete float type used as a converter. Its lookbehind lets a type parameter past
and its trailing identifier character lets a literal past, so `Vector{Float64}`,
`::Float64`, `Float64[]` and `Float64(1)` are all untouched. Run against `src/` and `ext/`
it found six sites. Five were redundant and are deleted rather than exempted: `new` already
converts into a `min_score::Float64` field, and an assignment into a `Matrix{Float64}`
already converts. The sixth is the one the issue names, `unify_gaps`, which ADR 0135 and
issue #988 own; it takes the census's first named exemption, and the census asserts each
exemption is still a site, so the entry reds the build when #988 lands and it is not
removed.

**Two more red files, both defects the toolchain only revealed.** `test_02` compared
`res.vars[idx]` against `literal[idx]` — one permutation applied to two different
orderings, which passed only while the term `Dict` iterated in the order the literal was
written. Each side sorts by its own names now. `test_09i`'s alias census read
`names(PO; all = true)`, which also answers the compiler's gensym bindings; a gensym is
bound to the very alias its keyword annotates, so it counted every alias twice. It skips
`Base.isgensym` now.

**Two Docs failures.** `risk_budget_constraints`'s doctest wrote an overlapping budget as
a plain `Dict`, so which of `"A"` and `"group1"` won was hash order — the very
non-determinism `estimator_to_val`'s own docstring warns about. It is a vector of pairs
now, and the recorded answer is the one it always meant. Separately,
`examples/3_optimisers/16_Feature_Distance_Clustering.jl` was the one example that read
`Factors.csv.gz` unsliced while its asset table took `[(end - 252):end]`; #985 stopped the
conversion imputing the outer join, so every asset carried a non-finite return at some
observation and the Coverage Universe was empty. It takes the same slice as its seven
siblings.

**The toolchain.** `setup-julia` with `version: "1"` means whatever Julia is newest, and on
2026-09-10 that became 1.13.0 — one release past `Project.toml`'s `julia = "1.12, 1.11"`.
`Test`, `TestOnPRs` and `Docs` pin `"1.12"`, joining the gates that already read their
version from `code_health/Manifest.toml`. Widening compat instead is #998's to decide, and
it costs a re-record of the four doctests that print a `UniverseSets.dict` and a full-suite
run nobody has done on 1.13.

Measured on 1.12.7: `test_02`, `test_09i`, `test_41`, `test_55`, `test_08a`, `test_08r`,
`test_08`, `test_25`, `test_26`, `test_45` and `test_54` pass; the full doctest run passes;
the example runs to its last table; `jet.jl`, `complexity.jl`, `expansion.jl`, `size.jl`
and `sweep_check.jl` are all green. The local 1.13 toolchain has a broken `Pkg` stdlib, so
nothing here is measured on 1.13.

+++ ext/PortfolioOptimisersPlotsExt.jl
@@ -2546,7 +2546,7 @@ function factor_summary_columns(fs::PortfolioOptimisers.FactorSummaryResult)
     for (r, i) in enumerate(keep)
         v = vals[i]
         for k in 1:K
-            M[r, k] = Float64(v[k])
+            M[r, k] = v[k]
         end
     end
     return M, FACTOR_SUMMARY_LABELS[keep], length(keep) < length(vals)
@@ -3084,7 +3084,7 @@ function forecast_summary_columns(fs::PortfolioOptimisers.ForecastSummaryResult)
     for r in eachindex(vals)
         v = vals[r]
         for k in 1:K
-            M[r, k] = Float64(v[k])
+            M[r, k] = v[k]
         end
     end
     return M, labels, partial
